+++ src/main/java/org/ojalgo/structure/Factory2D.java
@@ -23,7 +23,9 @@ package org.ojalgo.structure;
 
 import java.util.List;
 
+import org.ojalgo.function.FunctionSet;
 import org.ojalgo.function.NullaryFunction;
+import org.ojalgo.scalar.Scalar.Factory;
 
 public interface Factory2D<I extends Structure2D> extends FactorySupplement {
 
@@ -89,6 +91,24 @@ public interface Factory2D<I extends Structure2D> extends FactorySupplement {
             return this.makeSparse(shape.countRows(), shape.countColumns());
         }
 
+    }
+
+    default Factory1D<I> asFactory1D() {
+        return new Factory1D<I>() {
+
+            public FunctionSet<?> function() {
+                return Factory2D.this.function();
+            }
+
+            public I make(final long count) {
+                return Factory2D.this.make(count, 1L);
+            }
+
+            public Factory<?> scalar() {
+                return Factory2D.this.scalar();
+            }
+
+        };
     }
 
     I make(long rows, long columns);